# Optimize MessageFactory with compiled lambda expressions