--- a/Source/EasyNetQ.sln
+++ b/Source/EasyNetQ.sln
@@ -7,12 +7,12 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".Solution Items", ".Solutio
 	ProjectSection(SolutionItems) = preProject
 		..\.editorconfig = ..\.editorconfig
 		..\.gitignore = ..\.gitignore
+		..\.github\dependabot.yml = ..\.github\dependabot.yml
 		..\.github\workflows\ci.yml = ..\.github\workflows\ci.yml
 		..\.github\workflows\depsreview.yaml = ..\.github\workflows\depsreview.yaml
-		..\.github\dependabot.yml = ..\.github\dependabot.yml
 		Directory.Build.props = Directory.Build.props
-		FodyWeavers.xml = FodyWeavers.xml
 		docker-compose.yml = docker-compose.yml
+		FodyWeavers.xml = FodyWeavers.xml
 		FodyWeavers.xsd = FodyWeavers.xsd
 		..\hall_of_fame.md = ..\hall_of_fame.md
 		..\licence.txt = ..\licence.txt
@@ -63,7 +63,8 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EasyNetQ.Examples.DlxRetry", "EasyNetQ.Examples.DlxRetry\EasyNetQ.Examples.DlxRetry.csproj", "{334FC79A-41F7-4506-B887-8FD4ADCC28E4}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Examples", "Examples", "{C24EB2CE-59ED-4348-903F-2DF7E868E7BD}"
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EasyNetQ.Serialization.SystemTextJson", "EasyNetQ.Serialization.SystemTextJson\EasyNetQ.Serialization.SystemTextJson.csproj", "{5C190112-DD9E-451B-8BD8-C6F5BF88D4B1}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EasyNetQ.Serialization.SystemTextJson", "EasyNetQ.Serialization.SystemTextJson\EasyNetQ.Serialization.SystemTextJson.csproj", "{5C190112-DD9E-451B-8BD8-C6F5BF88D4B1}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EasyNetQ.Examples.PubSubWithAdvancedApi", "EasyNetQ.Examples.PubSubWithAdvancedApi\EasyNetQ.Examples.PubSubWithAdvancedApi.csproj", "{3033D84D-F003-4993-95E0-6898E487EF38}"
 EndProject
@@ -143,6 +144,10 @@ Global
 		{334FC79A-41F7-4506-B887-8FD4ADCC28E4}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{334FC79A-41F7-4506-B887-8FD4ADCC28E4}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{334FC79A-41F7-4506-B887-8FD4ADCC28E4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5C190112-DD9E-451B-8BD8-C6F5BF88D4B1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5C190112-DD9E-451B-8BD8-C6F5BF88D4B1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5C190112-DD9E-451B-8BD8-C6F5BF88D4B1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5C190112-DD9E-451B-8BD8-C6F5BF88D4B1}.Release|Any CPU.Build.0 = Release|Any CPU
 		{3033D84D-F003-4993-95E0-6898E487EF38}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{3033D84D-F003-4993-95E0-6898E487EF38}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3033D84D-F003-4993-95E0-6898E487EF38}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -151,10 +156,6 @@ Global
 		{34907240-61E8-425B-8444-3F5676F8F865}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{34907240-61E8-425B-8444-3F5676F8F865}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{34907240-61E8-425B-8444-3F5676F8F865}.Release|Any CPU.Build.0 = Release|Any CPU
-		{5C190112-DD9E-451B-8BD8-C6F5BF88D4B1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5C190112-DD9E-451B-8BD8-C6F5BF88D4B1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5C190112-DD9E-451B-8BD8-C6F5BF88D4B1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5C190112-DD9E-451B-8BD8-C6F5BF88D4B1}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -175,9 +176,9 @@ Global
 		{1CAE3C49-52AB-45D4-B0DD-784E7C870118} = {83745BDD-49A0-4247-A5AC-89428D7D0B7D}
 		{8A743708-92DE-4199-B3C9-A52D7F521C8D} = {83745BDD-49A0-4247-A5AC-89428D7D0B7D}
 		{334FC79A-41F7-4506-B887-8FD4ADCC28E4} = {C24EB2CE-59ED-4348-903F-2DF7E868E7BD}
+		{5C190112-DD9E-451B-8BD8-C6F5BF88D4B1} = {0BF00E29-CFEB-43C0-B8A7-98425AD7C303}
 		{3033D84D-F003-4993-95E0-6898E487EF38} = {C24EB2CE-59ED-4348-903F-2DF7E868E7BD}
 		{34907240-61E8-425B-8444-3F5676F8F865} = {C24EB2CE-59ED-4348-903F-2DF7E868E7BD}
-		{5C190112-DD9E-451B-8BD8-C6F5BF88D4B1} = {0BF00E29-CFEB-43C0-B8A7-98425AD7C303}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {A3DE3E7F-B61C-4398-AA54-A0DDE3BD850E}