--- a/Source/EasyNetQ/MessageFactory.cs
+++ b/Source/EasyNetQ/MessageFactory.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Concurrent;
+using System.Linq.Expressions;
 
 namespace EasyNetQ;
 
@@ -9,11 +10,29 @@ namespace EasyNetQ;
 /// </summary>
 public static class MessageFactory
 {
-    private static readonly ConcurrentDictionary<Type, Type> GenericMessageTypesMap = new();
+    private static readonly ConcurrentDictionary<Type, Func<object?, MessageProperties, IMessage>> InstanceActivators = new();
 
     public static IMessage CreateInstance(Type messageType, object? body, MessageProperties properties)
     {
-        var genericType = GenericMessageTypesMap.GetOrAdd(messageType, t => typeof(Message<>).MakeGenericType(t));
-        return (IMessage)Activator.CreateInstance(genericType, body, properties)!;
+        var activator = InstanceActivators.GetOrAdd(messageType, bodyType =>
+        {
+            var genericMessageType = typeof(Message<>).MakeGenericType(bodyType);
+
+            var constructor = genericMessageType.GetConstructor(new[] { bodyType, typeof(MessageProperties) })!;
+            var bodyParameter = Expression.Parameter(typeof(object));
+            var propertiesParameter = Expression.Parameter(typeof(MessageProperties));
+
+            Expression<Func<object?, MessageProperties, IMessage>> expression =
+                Expression.Lambda<Func<object?, MessageProperties, IMessage>>(
+                    Expression.New(
+                        constructor,
+                        Expression.Convert(bodyParameter, bodyType),
+                        Expression.Convert(propertiesParameter, typeof(MessageProperties))),
+                    bodyParameter, propertiesParameter);
+
+            return expression.Compile();
+        });
+
+        return activator(body, properties);
     }
 }