--- a/hall_of_fame.md
+++ b/hall_of_fame.md
@@ -70,3 +70,4 @@ No particular order. Don't forget to add your name with your pull request.
 * Ivan Maximov
 * Jens Willmer
 * Pierre Malatre
+* Piotr Zięzio