In new .NET versions, especially since NET 5.0 and newer, compiled lambda expressions are much faster than `Activator.CreateInstance` when creating new objects dynamically.

I'd like to share an implementation of MessageFactory that we're using. It is about 20 times faster than the current one.
When message count is in range of 10k/s just creating new message instances takes time in order of single milliseconds, which I find significant enough.
I didn't add any input parameter validation for performance reasons so it will throw in the same cases as existing implementation.
There already is a test checking that it is working correctly (`MessageFactoryTests.Should_correctly_create_generic_message`).

Quick benchmark for comparison: https://github.com/pziezio/EasyNetQ/commit/ceee32d8b449ae98b5db0a3075937788c212d9f2
There are probably more things worth benchmarking in the library so it could be a good starting point.

```
BenchmarkDotNet=v0.13.2, OS=Windows 10 (10.0.19044.2130/21H2/November2021Update)
Intel Core i7-9700KF CPU 3.60GHz (Coffee Lake), 1 CPU, 8 logical and 8 physical cores
.NET SDK=6.0.402
  [Host]     : .NET 6.0.10 (6.0.1022.47605), X64 RyuJIT AVX2
  DefaultJob : .NET 6.0.10 (6.0.1022.47605), X64 RyuJIT AVX2

|                       Method |       N |       Mean |     Error |    StdDev | Ratio | RatioSD |
|----------------------------- |-------- |-----------:|----------:|----------:|------:|--------:|
|              NewOperatorTest | 1000000 |   8.958 ns | 0.2432 ns | 0.2703 ns |  1.00 |    0.00 |
| ExpressionMessageFactoryTest | 1000000 |  20.951 ns | 0.1703 ns | 0.1593 ns |  2.32 |    0.07 |
|  ActivatorMessageFactoryTest | 1000000 | 396.482 ns | 6.8984 ns | 6.4527 ns | 43.97 |    1.57 |
```

Please feel free to review this. I am not 100% sure if this will work on all architectures and operating systems because I cannot check all cases. But since it can be compiled as a netstandard2.0 library I think it will work everywhere.

BTW `EndProject` tag after `EasyNetQ.Serialization.SystemTextJson` is missing in the solution file.
It was changed recently in https://github.com/EasyNetQ/EasyNetQ/commit/47b5f5df51de6cae94951de84e53dd5ce6b97f09#diff-6b88f86d05c92e346ef4aaeaf3068b4a81a6d63830ade56998ff09f3058f9ec0