@sungam3r 2022-11-09
I would support both cases - when runtime supports code compilation at runtime and not. Check this https://stackoverflow.com/questions/58599459/how-can-i-detect-runtime-code-generation-in-net-core 

From MS sources:
```c#
public static ServiceProvider BuildServiceProvider(this IServiceCollection services, ServiceProviderOptions options)
{
            if (services == null)
            {
                throw new ArgumentNullException(nameof(services));
            }

            if (options == null)
            {
                throw new ArgumentNullException(nameof(options));
            }

            IServiceProviderEngine engine;

#if !NETCOREAPP
            engine = new DynamicServiceProviderEngine(services);
#else
            if (RuntimeFeature.IsDynamicCodeCompiled)
            {
                engine = new DynamicServiceProviderEngine(services);
            }
            else
            {
                // Don't try to compile Expressions/IL if they are going to get interpreted
                engine = new RuntimeServiceProviderEngine(services);
            }
#endif

            return new ServiceProvider(services, engine, options);
        }

```