@Pliner 2022-11-09
@sungam3r Interesting, but I suppose we should investigate how to change all existing code, because expression trees are used a lot, but the implementation is fine at the moment.