@pziezio 2022-11-09
@sungam3r MSDN says that `RuntimeFeature.IsDynamicCodeCompiled` is supported from `netstandard2.1`. EasyNetQ is currently at `netstandard2.0`.
What's more, `MessageFactory` is a simple static class. Checking for dynamic code support at every message creation would probably cancel all performance gain. We can extract `IMessageFactory` to an interface and dynamically register supported implementation in the container. But it's still one additional redirection compared to now.

I will make some additional benchmarking for that.