@sungam3r 2022-11-10
> MSDN says that RuntimeFeature.IsDynamicCodeCompiled is supported from netstandard2.1. EasyNetQ is currently at netstandard2.0.

Yes.

> Checking for dynamic code support at every message creation would probably cancel all performance gain. We can extract IMessageFactory to an interface and dynamically register supported implementation in the container. 

I'm fine to merge as is. I only spotted your attention at this capability. For someone dynamic code compilation **may** be a problem but not to me.